timestamping: WIP
work in progress. Adding support for timestamping in BSD Sockets API.
This can be accomplished by adding SO_TIMESTAMPING option with
setsockopt(), it's a socket level option that accepts bitfields instead
of boolean.

Started with TX timestamping that seems a little bit harder.
In other implementation TX timestamps are retreived from ERR_QUEUE,
but here I wanted to do it in different way as there is no ERR_QUEUE
for the socket.

Other solution that seems easier at that stage is to use callbacks
for timestamp generation event. The mechanism is ready so I'll move
forvard with that and leave this one.

Signed-off-by: Adam Wojasinski <awojasinski@baylibre.com>